Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
opencv-python
ultralytics
alembic
python-dotenv
